polish name
add more include

make StreamWrapper into tensor.h

remove additional dtype

make enum as enum class

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

debug

remove additional tests
